rag_app/requirements.txt
boto3
python-dotenv
google-generativeai
opensearch-py
requests-aws4auth
